test: Add test to lex operator tokens
Fixed a bug hidden within the compare_sequence function that was
comparing the length of the sequence with itself instead of the best
sequence. Also fixed bugs where tokens were being read in cases that
were returning errors. We don't want tokens read in those cases because
it will increase the position of the mark. To mitigate any of those
bugs from occuring in the future I added a check that will overwrite
the `best_token` if the best token is an error and the new token isn't
regardless of the marks seek position.